+++ projects/CoinFlip/CoinFlip.Model/src/Balance/IBalance.cs
@@ -7,6 +7,10 @@ namespace CoinFlip
 	public interface IBalance :INotifyPropertyChanged
 	{
 		// Notes:
+		//  - Doubles vs. Decimals:
+		//    Decimals should only be used for storing the floating-point numbers and simple arithmetic operations on them.
+		//    All heavy mathematics like calculating indicators for technical analysis should be performed on Double values.
+		//    So, using decimal for balances, double for charts etc.
 		//  - This interface is used to isolate the rest of the code from the
 		//    Fund/Balance system (which is a bit confusing).
 
@@ -20,13 +24,13 @@ namespace CoinFlip
 		DateTimeOffset LastUpdated { get; }
 
 		/// <summary>The total amount of the currency 'Coin' associated with this fund on 'Coin.Exchange'</summary>
-		Unit<double> Total { get; }
+		Unit<decimal> Total { get; }
 
 		/// <summary>Total amount able to be used for new trades</summary>
-		Unit<double> Available { get; }
+		Unit<decimal> Available { get; }
 
 		/// <summary>Total amount set aside for pending orders and trade strategies</summary>
-		Unit<double> Held { get; }
+		Unit<decimal> Held { get; }
 
 		/// <summary>Holds on this fund balance</summary>
 		FundHoldContainer Holds { get; }
